Refactored codebase to src directory
